# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lead Tracker API",
    "description": "Local business prospecting: search, audit, score, call, export.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get App Settings",
        "operationId": "get_app_settings_api_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsOut"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "settings"
        ],
        "summary": "Patch App Settings",
        "operationId": "patch_app_settings_api_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "object",
                "title": "Patch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/presets": {
      "get": {
        "tags": [
          "presets"
        ],
        "summary": "List Presets",
        "operationId": "list_presets_api_presets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PresetOut"
                  },
                  "type": "array",
                  "title": "Response List Presets Api Presets Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "presets"
        ],
        "summary": "Create Preset",
        "operationId": "create_preset_api_presets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/presets/{preset_id}": {
      "patch": {
        "tags": [
          "presets"
        ],
        "summary": "Update Preset",
        "operationId": "update_preset_api_presets__preset_id__patch",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Preset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "presets"
        ],
        "summary": "Delete Preset",
        "operationId": "delete_preset_api_presets__preset_id__delete",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Preset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/suppression": {
      "get": {
        "tags": [
          "suppression"
        ],
        "summary": "List Suppression",
        "operationId": "list_suppression_api_suppression_get",
        "parameters": [
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inactive"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SuppressionOut"
                  },
                  "title": "Response List Suppression Api Suppression Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "suppression"
        ],
        "summary": "Add Suppression",
        "operationId": "add_suppression_api_suppression_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuppressionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuppressionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/suppression/{entry_id}": {
      "delete": {
        "tags": [
          "suppression"
        ],
        "summary": "Reenable Suppression",
        "description": "Explicitly re-enable a suppressed number (the entry is kept, deactivated).",
        "operationId": "reenable_suppression_api_suppression__entry_id__delete",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuppressionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CsvSettings": {
        "properties": {
          "delimiter": {
            "type": "string",
            "enum": [
              ",",
              ";",
              "\t"
            ],
            "title": "Delimiter",
            "default": ","
          },
          "include_bom": {
            "type": "boolean",
            "title": "Include Bom",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "delimiter",
          "include_bom"
        ],
        "title": "CsvSettings"
      },
...
      "EnvironmentInfo": {
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/ProviderInfo"
          },
          "provider_api_key_set": {
            "type": "boolean",
            "title": "Provider Api Key Set"
          },
          "pagespeed_api_key_set": {
            "type": "boolean",
            "title": "Pagespeed Api Key Set"
          },
          "places_field_tier": {
            "type": "string",
            "title": "Places Field Tier"
          },
          "demo_mode": {
            "type": "boolean",
            "title": "Demo Mode"
          },
          "auth_mode": {
            "type": "string",
            "title": "Auth Mode"
          },
          "environment": {
            "type": "string",
            "title": "Environment"
          },
          "google_latlng_retention_days": {
            "type": "integer",
            "title": "Google Latlng Retention Days"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "provider_api_key_set",
          "pagespeed_api_key_set",
          "places_field_tier",
          "demo_mode",
          "auth_mode",
          "environment",
          "google_latlng_retention_days"
        ],
        "title": "EnvironmentInfo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "OkResponse"
      },
...
      "OpportunityType": {
        "type": "string",
        "enum": [
          "NO_WEBSITE",
          "BROKEN_WEBSITE",
          "OUTDATED_WEBSITE",
          "WEBSITE_REDESIGN",
          "MOBILE_PROBLEM",
          "NO_BOOKING",
          "NO_CONTACT_CTA",
          "WEAK_CONVERSION",
          "SLOW_WEBSITE",
          "GOOGLE_PROFILE",
          "LOW_REVIEWS",
          "MISSING_BUSINESS_INFO"
        ],
        "title": "OpportunityType"
      },
...
      "PresetCreate": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Label"
          },
          "label_bg": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Label Bg"
          },
          "category_query": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Category Query"
          },
          "included_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Included Type"
          },
          "keywords": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          },
          "match_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Match Types"
          },
          "calling_window_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calling Window Start"
          },
          "calling_window_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calling Window End"
          },
          "booking_oriented": {
            "type": "boolean",
            "title": "Booking Oriented",
            "default": false
          },
          "discovery_dependent": {
            "type": "boolean",
            "title": "Discovery Dependent",
            "default": true
          },
          "strong_opportunities": {
            "items": {
              "$ref": "#/components/schemas/OpportunityType"
            },
            "type": "array",
            "title": "Strong Opportunities"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 100
          },
          "key": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "pattern": "^[a-z0-9_]+$",
            "title": "Key"
          }
        },
        "type": "object",
        "required": [
          "label",
          "category_query",
          "key"
        ],
        "title": "PresetCreate"
      },
      "PresetOut": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Label"
          },
          "label_bg": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Label Bg"
          },
          "category_query": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Category Query"
          },
          "included_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Included Type"
          },
          "keywords": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          },
          "match_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Match Types"
          },
          "calling_window_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calling Window Start"
          },
          "calling_window_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calling Window End"
          },
          "booking_oriented": {
            "type": "boolean",
            "title": "Booking Oriented",
            "default": false
          },
          "discovery_dependent": {
            "type": "boolean",
            "title": "Discovery Dependent",
            "default": true
          },
          "strong_opportunities": {
            "items": {
              "$ref": "#/components/schemas/OpportunityType"
            },
            "type": "array",
            "title": "Strong Opportunities"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 100
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "is_builtin": {
            "type": "boolean",
            "title": "Is Builtin"
          },
          "lead_count": {
            "type": "integer",
            "title": "Lead Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "label",
          "label_bg",
          "category_query",
          "included_type",
          "keywords",
          "match_types",
          "calling_window_start",
          "calling_window_end",
          "booking_oriented",
          "discovery_dependent",
          "strong_opportunities",
          "sort_order",
          "id",
          "key",
          "is_builtin",
          "lead_count"
        ],
        "title": "PresetOut"
      },
      "PresetUpdate": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "label_bg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label Bg"
          },
          "category_query": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Query"
          },
          "included_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Included Type"
          },
          "keywords": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          },
          "match_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Types"
          },
          "calling_window_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calling Window Start"
          },
          "calling_window_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calling Window End"
          },
          "booking_oriented": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booking Oriented"
          },
          "discovery_dependent": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discovery Dependent"
          },
          "strong_opportunities": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OpportunityType"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strong Opportunities"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order"
          }
        },
        "type": "object",
        "title": "PresetUpdate"
      },
...
      "RuntimeSettings": {
        "properties": {
          "general": {
            "$ref": "#/components/schemas/GeneralSettings"
          },
          "calling": {
            "$ref": "#/components/schemas/CallingSettings"
          },
          "audit": {
            "$ref": "#/components/schemas/AuditSettings"
          },
          "search": {
            "$ref": "#/components/schemas/SearchSettings"
          },
          "csv": {
            "$ref": "#/components/schemas/CsvSettings"
          },
          "pricing": {
            "$ref": "#/components/schemas/PricingSettings"
          },
          "scoring": {
            "$ref": "#/components/schemas/ScoringConfig"
          }
        },
        "type": "object",
        "required": [
          "general",
          "calling",
          "audit",
          "search",
          "csv",
          "pricing",
          "scoring"
        ],
        "title": "RuntimeSettings"
      },
...
      "SettingsOut": {
        "properties": {
          "runtime": {
            "$ref": "#/components/schemas/RuntimeSettings"
          },
          "environment": {
            "$ref": "#/components/schemas/EnvironmentInfo"
          }
        },
        "type": "object",
        "required": [
          "runtime",
          "environment"
        ],
        "title": "SettingsOut"
      },
...
      "SuppressionCreate": {
        "properties": {
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "business_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "SuppressionCreate"
      },
      "SuppressionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "normalized_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Phone"
          },
          "business_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "created_by_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "deactivated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deactivated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "normalized_phone",
          "business_id",
          "reason",
          "active",
          "created_by_id",
          "created_at",
          "deactivated_at"
        ],
        "title": "SuppressionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}